Add files via upload
The aim of this mini ORM is to manage the access of a table in database from separate layer written in java. The most benefit from this is that it simplifies database access operations through some methods like persist() and find(). This layer handles all database related queries inside it. My goal with this is to receive a better understanding of how ORMs like Hibernate work.

Functionalities for creating table if not exist and altering table need to be added!! Also the code need some refactoring!!